# Improve offset reading performance