Simply change multiple(3 * filenum) calls to 1 call. This results in roughly 90ms reading offsets(170ms total for list) to 1.5ms(80ms) on test machine while listing TR_Data.bsa