@arviceblot 2023-08-16
Looks good, thanks for your contribution! I will admit the implementation of this tool is a bit naive as it was my first foray into rust.